Create utility queryAs method in DatabaseUtils
