+++ app/src/main/java/com/github/geohunt/app/utility/TaskUtils.kt
@@ -3,7 +3,6 @@ package com.github.geohunt.app.utility
 import android.app.Activity
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
-import com.google.firebase.database.DataSnapshot
 import java.util.concurrent.CompletableFuture
 import kotlin.coroutines.cancellation.CancellationException
 
@@ -26,22 +25,6 @@ fun <TResult> Task<TResult>.toCompletableFuture(activity: Activity) : Completabl
             completableFuture.completeExceptionally(CancellationException())
         }
     return completableFuture
-}
-
-/**
- * Converts the given [DataSnapshot] from Firebase to a map.
- *
- * @return The map, where the keys present in the map are mapped to their corresponding Firebase values.
- */
-inline fun <reified T> DataSnapshot.toMap(): Map<String, T>
-{
-    val map = mutableMapOf<String, T>()
-    for (child in this.children) {
-        if (child.value is T) {
-            map[child.key!!] = child.value as T
-        }
-    }
-    return map
 }
 
 /**
